link nmcli steps into nmtui
nmtui was keeping its own version of some steps, which is hard to
maintain. Now, all nmcli steps can be used in nmtui, they are
"symlinked".

Noticing different timeout in some steps, additional_sleep function wass
added on this purpose. This function has different definitions in nmcli
and nmtui. This concept can be used also in another scenarios, where
nmcli and nmtui version of steps should differ for some reason.
