+++ nmcli/features/steps/bond_bridge_team.py
@@ -9,7 +9,7 @@ import subprocess
 from subprocess import Popen, check_output, call
 from glob import glob
 
-from steps import run, command_output, command_code
+from steps import command_output, command_code, additional_sleep
 
 
 
@@ -149,6 +149,15 @@ def external_bridge_check(context, number):
 
 @step(u'Team "{team}" is down')
 def team_is_down(context, team):
+    additional_sleep(2)
     cmd = pexpect.spawn('teamdctl %s state dump' %team, logfile=context.log, encoding='utf-8')
     print (command_code(context, 'teamdctl %s state dump' %team))
     assert command_code(context, 'teamdctl %s state dump' %team) != 0, 'team "%s" exists' % (team)
+
+
+@step(u'Team "{team}" is up')
+def team_is_down(context, team):
+    additional_sleep(2)
+    cmd = pexpect.spawn('teamdctl %s state dump' %team, logfile=context.log, encoding='utf-8')
+    print (command_code(context, 'teamdctl %s state dump' %team))
+    assert command_code(context, 'teamdctl %s state dump' %team) == 0, 'team "%s" does not exist' % (team)
