+++ nmcli/features/steps/connection.py
@@ -9,7 +9,7 @@ import subprocess
 from subprocess import Popen, check_output, call
 from glob import glob
 
-from steps import run, command_output, command_code
+from steps import command_output, command_code, additional_sleep
 
 
 
@@ -143,6 +143,12 @@ def bring_down_connection_ignoring(context, connection):
     r = cli.expect([pexpect.EOF, pexpect.TIMEOUT])
     if r == 1:
         raise Exception('nmcli connection down %s timed out (180s)' % connection)
+
+
+@step(u'Bring up connection "{connection}" ignoring everything')
+def bring_up_connection_ignore_everything(context, connection):
+    subprocess.Popen('nmcli connection up %s' % connection, shell=True)
+    sleep(1)
 
 
 @step(u'Check if "{name}" is active connection')
