+++ nmtui/features/steps/commands.py
@@ -1,0 +1,1 @@
+../../../nmcli/features/steps/commands.py
