+++ nmtui/features/steps/device.py
@@ -1,0 +1,1 @@
+../../../nmcli/features/steps/device.py
