+++ nmtui/features/steps/editor.py
@@ -1,0 +1,1 @@
+../../../nmcli/features/steps/editor.py
