+++ nmtui/features/steps/vpn.py
@@ -1,0 +1,1 @@
+../../../nmcli/features/steps/vpn.py
